+++ packages/implementations/react/src/atomic/Toggle/src/toggle.stories.tsx
@@ -1,6 +1,10 @@
+import React from "react";
 import type { Meta, StoryObj } from "@storybook/react";
 import { argTypesConvert } from ".storybook/utils";
+import { Icon } from "@rarui-react/icon";
+import { LightbulbFilledIcon, LightbulbOutlinedIcon } from "@rarui/icons";
 import { Toggle } from "./Toggle";
+
 import docs from "./toggle.docs.json";
 
 const meta: Meta<typeof Toggle> = {
@@ -14,7 +18,16 @@ export default meta;
 type Story = StoryObj<typeof Toggle>;
 
 export const basic: Story = {
-  args: {},
+  args: {
+    icons: {
+      unchecked: (
+        <Icon source={<LightbulbOutlinedIcon size="small" />} color="$invert" />
+      ),
+      checked: (
+        <Icon source={<LightbulbFilledIcon size="small" />} color="$primary" />
+      ),
+    },
+  },
 };
 
 export const error: Story = {